fix: null Inn employment error in deal db

+++ application/src/main/java/com/application/model/dto/LoanOfferDTO.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @Schema(title = "Loan Offer", description = "Loan offer after short scoring.")
@@ -17,12 +19,12 @@ public class LoanOfferDTO {
     @Schema(description = "Requested loan amount.", example = "1000000.00", minimum = "10000.00")
     @NotNull
     @Min(10000)
-    private Double requestedAmount;
+    private BigDecimal requestedAmount;
 
     @Schema(description = "Total loan amount (with services, insurance, etc.).", example = "1000000.00", minimum = "10000.00")
     @NotNull
     @Min(10000)
-    private Double totalAmount;
+    private BigDecimal totalAmount;
 
     @Schema(description = "Requested loan term (months).", example = "24", minimum = "6")
     @NotNull
@@ -31,11 +33,11 @@ public class LoanOfferDTO {
 
     @Schema(description = "Monthly payment.", example = "10000.10")
     @NotNull
-    private Double monthlyPayment;
+    private BigDecimal monthlyPayment;
 
     @Schema(description = "Loan rate", example = "12.50")
     @NotNull
-    private Double rate;
+    private BigDecimal rate;
 
     @Schema(description = "Is insurance enabled?", example = "true")
     @NotNull
